[workflow] Split step_log_types into separate header to fix AUTOMOC
Extract step_log_level and step_log_entry into step_log_types.hpp, which
has no rfl.hpp dependency. Qt AUTOMOC's simplified C++ parser cannot handle
reflect-cpp templates; including rfl.hpp transitively via workflow_events.hpp
caused it to silently skip moc_WorkflowStepLogWidget.cpp, leaving
staticMetaObject undefined at runtime.

workflow_events.hpp now includes step_log_types.hpp (keeping the rfl
Reflector specialisations), and workflow_query_protocol.hpp and
WorkflowStepLogWidget.hpp switch to the lighter header.
