+++ projects/ores.qt.workflow/include/ores.qt/WorkflowStepLogWidget.hpp
@@ -24,7 +24,7 @@
 #include <QWidget>
 #include <QTableWidget>
 #include "ores.qt/WorkflowExport.hpp"
-#include "ores.workflow.api/messaging/workflow_events.hpp"
+#include "ores.workflow.api/messaging/step_log_types.hpp"
 
 namespace ores::qt {
